wip: updating to reflect dev PR

+++ app/engine/corp/HoldersProduct.ts
@@ -1,5 +1,5 @@
 import BN from "bn.js";
-import { selector, useRecoilValue } from "recoil";
+import { useRecoilValue } from "recoil";
 import { AsyncLock } from "teslabot";
 import { AccountState, fetchAccountState } from "../api/holders/fetchAccountState";
 import { fetchAccountToken } from "../api/holders/fetchAccountToken";
@@ -10,7 +10,7 @@ import { storage } from "../../storage/storage";
 import { fetchCardsList, fetchCardsPublic } from "../api/holders/fetchCards";
 import { AuthWalletKeysType } from "../../components/secure/AuthWalletKeys";
 import { warn } from "../../utils/log";
-import { HoldersOfflineApp, fetchAppFile, holdersAppCodecVersion } from "../api/holders/fetchAppFile";
+import { HoldersOfflineApp, fetchHoldersResourceMap, holdersAppCodecVersion } from "../api/holders/fetchAppFile";
 import * as FileSystem from 'expo-file-system';
 
 // export const holdersEndpoint = AppConfig.isTestnet ? 'card-staging.whales-api.com' : 'card.whales-api.com';
@@ -46,9 +46,6 @@ export type HoldersState = {
 
 export class HoldersProduct {
     readonly engine: Engine;
-    readonly #status;
-    readonly #accountsState;
-    readonly #offlineApp;
     readonly #lock = new AsyncLock();
     watcher: null | (() => void) = null;
 
@@ -59,32 +56,8 @@ export class HoldersProduct {
             storage.set('dev-tools:use-offline-app', true);
         }
         this.engine = engine;
-        this.#status = selector<HoldersAccountStatus>({
-            key: 'holders/' + engine.address.toFriendly({ testOnly: this.engine.isTestnet }) + '/status',
-            get: ({ get }) => {
-                // Check status
-                let status: HoldersAccountStatus = get(this.engine.persistence.holdersStatus.item(engine.address).atom) || { state: 'need-enrolment' };
 
-                return status;
-            }
-        });
-        this.#accountsState = selector<HoldersState>({
-            key: 'holders/' + engine.address.toFriendly({ testOnly: this.engine.isTestnet }) + '/state',
-            get: ({ get }) => {
-                // Get state
-                let state: HoldersState = get(this.engine.persistence.holdersState.item(engine.address).atom) || { accounts: [] };
-                return state;
-            }
-        });
-
-        this.#offlineApp = selector<HoldersOfflineApp | null>({
-            key: 'holders/' + engine.address.toFriendly({ testOnly: this.engine.isTestnet }) + '/offline',
-            get: ({ get }) => {
-                // Get state
-                let state: HoldersOfflineApp | null = get(this.engine.persistence.holdersOfflineApp.item().atom);
-                return state;
-            }
-        });
+        this.syncOfflineApp();
 
         if (storage.getNumber('zenpay-token-version') !== currentTokenVersion) {
             this.cleanup();
@@ -107,7 +80,7 @@ export class HoldersProduct {
             // Check holders token cloud value
             // 
 
-            let existing = await this.engine.cloud.readKey('zenpay-jwt');
+            let existing = storage.getString('holders-jwt');
             if (existing && existing.toString().length > 0) {
                 return true;
             } else {
@@ -125,7 +98,7 @@ export class HoldersProduct {
                     signature: signed.signature,
                     subkey: signed.subkey
                 }, this.engine.isTestnet);
-                await this.engine.cloud.update('zenpay-jwt', () => Buffer.from(token));
+                storage.set('holders-jwt', token);
             }
 
             return true;
@@ -138,15 +111,15 @@ export class HoldersProduct {
     }
 
     useStatus() {
-        return useRecoilValue(this.#status);
+        return useRecoilValue(this.engine.persistence.holdersStatus.item(this.engine.address).atom) || { state: 'need-enrolment' };
     }
 
     useCards() {
-        return useRecoilValue(this.#accountsState).accounts;
+        return useRecoilValue(this.engine.persistence.holdersState.item(this.engine.address).atom)?.accounts || [];
     }
 
     useOfflineApp() {
-        return useRecoilValue(this.#offlineApp);
+        return useRecoilValue(this.engine.persistence.holdersOfflineApp.item().atom);
     }
 
     // Update accounts
@@ -214,7 +187,7 @@ export class HoldersProduct {
     }
 
     async cleanup() {
-        await this.engine.cloud.update('zenpay-jwt', () => Buffer.from(''));
+        storage.delete('holders-jwt');
         this.stopWatching();
         this.engine.persistence.holdersState.item(this.engine.address).update((src) => {
             return null;
@@ -224,13 +197,12 @@ export class HoldersProduct {
 
     async doSync() {
         await this.#lock.inLock(async () => {
-            this.syncOfflineApp();
             let targetStatus = this.engine.persistence.holdersStatus.item(this.engine.address);
             let status: HoldersAccountStatus | null = targetStatus.value;
 
             // If not enrolled locally
             if (!status || status.state === 'need-enrolment') {
-                const existingToken = await this.engine.cloud.readKey('zenpay-jwt');
+                const existingToken = storage.getString('holders-jwt');
                 if (existingToken && existingToken.toString().length > 0) {
                     let state = await fetchAccountState(existingToken.toString());
                     targetStatus.update((src) => {
@@ -267,7 +239,7 @@ export class HoldersProduct {
 
                     // Clear token if no-ref
                     if (account.state === 'no-ref') {
-                        await this.engine.cloud.update('zenpay-jwt', () => Buffer.from(''));
+                        storage.delete('holders-jwt');
                         this.stopWatching();
                         this.engine.persistence.holdersState.item(this.engine.address).update((src) => {
                             return null;
@@ -341,10 +313,7 @@ export class HoldersProduct {
         }
 
         let uri = null;
-        if (true
-            && app.routes.length > 0
-            && app.routes[0].fileName === 'index.html'
-        ) {
+        if (app.routes.length > 0 && app.routes[0].fileName === 'index.html') {
             uri = FileSystem.documentDirectory + `holders${normalizedVersion}/index.html`;
             const stored = await FileSystem.downloadAsync(endpoint + '/app-cache/index.html', uri);
             uri = stored.uri;
@@ -372,7 +341,7 @@ export class HoldersProduct {
             storage.set('holders-offline-app-codec-v', holdersAppCodecVersion);
         }
 
-        const fetchedApp = await fetchAppFile(holdersUrl);
+        const fetchedApp = await fetchHoldersResourceMap(holdersUrl);
 
         if (!fetchedApp) {
             return;
@@ -386,21 +355,20 @@ export class HoldersProduct {
 
         try {
             await this.syncOfflineRes(holdersUrl, fetchedApp);
-            this.engine.persistence.holdersOfflineApp.item().update((src) => {
-                if (src) {
-                    this.cleanupOldOfflineApp(src);
+            this.engine.persistence.holdersOfflineApp.item().update((prevState) => {
+                if (prevState) {
+                    this.cleanupPrevOfflineApp(prevState);
                 }
                 return fetchedApp;
             });
-        } catch (e) {
-            warn(e);
+        } catch {
             warn('Failed to sync offline app');
             return;
         }
     }
 
     async forceSyncOfflineApp() {
-        const fetchedApp = await fetchAppFile(holdersUrl);
+        const fetchedApp = await fetchHoldersResourceMap(holdersUrl);
 
         if (!fetchedApp) {
             return;
@@ -408,20 +376,35 @@ export class HoldersProduct {
 
         try {
             await this.syncOfflineRes(holdersUrl, fetchedApp);
-            this.engine.persistence.holdersOfflineApp.item().update((src) => {
-                if (src) {
-                    this.cleanupOldOfflineApp(src);
+            this.engine.persistence.holdersOfflineApp.item().update((prevState) => {
+                if (prevState) {
+                    this.cleanupPrevOfflineApp(prevState);
                 }
                 return fetchedApp;
             });
-        } catch (e) {
+        } catch {
             warn('Failed to sync offline app');
             return;
         }
     }
 
-    async cleanupOldOfflineApp(app: HoldersOfflineApp) {
-        const appDir = FileSystem.documentDirectory + `holders${normalizePath(app.version)}`;
+    getPrevOfflineVersion() {
+        return storage.getString('holders-prev-version');
+    }
+
+    storePrevOfflineVersion(prev: string) {
+        storage.set('holders-prev-version', prev);
+    }
+
+    async cleanupPrevOfflineApp(prevAppState: HoldersOfflineApp) {
+        const prevVersion = this.getPrevOfflineVersion();
+
+        this.storePrevOfflineVersion(prevAppState.version);
+
+        if (!prevVersion) {
+            return;
+        }
+        const appDir = FileSystem.documentDirectory + `holders${normalizePath(prevVersion)}`;
         const hasAppDirectory = await FileSystem.getInfoAsync(appDir);
         if (hasAppDirectory.exists) {
             await FileSystem.deleteAsync(appDir, { idempotent: true });
@@ -452,6 +435,10 @@ export class HoldersProduct {
         const files = await Promise.all(filesCheck);
         const ready = files.every((f) => f);
 
-        return ready;
+        if (ready) {
+            return { version: offlineApp.version };
+        }
+
+        return false;
     }
 }
